Fix usage of factory and strategy pattern
