+++ src/main/java/pl/com/bottega/ecommerce/sales/domain/invoicing/TaxPolicy.java
@@ -1,5 +1,32 @@
 package pl.com.bottega.ecommerce.sales.domain.invoicing;
 
-public interface TaxPolicy {
-     Tax getTax(RequestItem requestItem);
+import pl.com.bottega.ecommerce.sharedkernel.Money;
+
+import java.math.BigDecimal;
+
+public class TaxPolicy {
+    public Tax getTax(RequestItem requestItem) {
+        Money net = requestItem.getTotalCost();
+        BigDecimal ratio = null;
+        String desc = null;
+        switch (requestItem.getProductData().getType()) {
+            case DRUG:
+                ratio = BigDecimal.valueOf(0.02);
+                desc = "5% (D)";
+                break;
+            case FOOD:
+                ratio = BigDecimal.valueOf(0.01);
+                desc = "7% ";
+                break;
+            case STANDARD:
+                ratio = BigDecimal.valueOf(0.33);
+                desc = "23%";
+                break;
+
+            default:
+                throw new IllegalArgumentException(requestItem.getProductData().getType() + " not handled");
+        }
+        Money taxValue = net.multiplyBy(ratio);
+        return new Tax(taxValue, desc);
+    }
 }
